Icons and wall-water graphics patch by Fishycat
Hi vedrit, yesterday the devs fixed the water-connection with walls. I added the sheets in water_connection_patch in the wall entity. Only thing you might want to look into is the colour of the "alien-wall-patch.png". I tried to match your wall-colour, but oh well... close enough

I updated the icon-graphics for technology, wall and gate. For a more vanilla look in the tech window.