Gave up on migration script, re-added minor global table data storage
The `game` table is unavailable during `on_load`, which means `update_walls()` and `update_current_tier` can't be called during it. So we have to go back to relying on storing tier and regen data in the `global` table. But in a much more clean way, copying it to a local variable for most of the time.

And the `global` table is unavailable inside a migration script, meaning the `global.alienwall` table is nil, so `update_walls()` doesn't work there either. Which leaves nothing for a migration script to actually do.

So we're pretty much stuck having to research a hybrid upgrade tech to get wall types to fix themselves when first upgrading to 0.7.0 in an existing save.

Everything else works great though!