A different sort of migration script
Seeing if flipping the upgrade researches off and on will trigger the wall tiers to update. Almost certainly not, as seen in the comments, but it's harmless.

The names don't need to be changed from the old to the new names, as that's handled by the .json file before this.